app/requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
serverless-wsgi==1.7.6
tensorflow==2.7.2
pandas>=1.3.0
pyjwt